Feature/RI-1541 finalizing create order

Approved-by: Calvin Usiri
Approved-by: Rutale Ivan

+++ app/src/main/java/io/ramani/ramaniStationary/data/home/models/response/DailySalesStatsRemoteModel.kt
@@ -8,7 +8,7 @@ data class DailySalesStatsRemoteModel(
     @SerializedName("date")
     val date: String?,
     @SerializedName("totalSales")
-    val totalSales: Int?,
+    val totalSales: Double?,
     @SerializedName("totalOrders")
     val totalOrders: Int?,
     @SerializedName("totalCanceledOrders")
